+++ tests/conftest.py
@@ -1,0 +1,18 @@
+"""shared fixtures"""
+
+from collections.abc import Generator
+from os import chdir
+from pathlib import Path
+from typing import Any
+
+import pytest
+
+
+@pytest.fixture
+def new_dir(tmp_path: Path) -> Generator[Path, Any, None]:
+    """Provide new test directory"""
+    current_directory = Path.cwd()
+    chdir(tmp_path)
+    yield tmp_path
+    # switch back to original directory
+    chdir(current_directory)