+++ tests/test_extend_gitlab_link.py
@@ -1,24 +1,9 @@
 """Test extend gitlab-link command"""
 
-from collections.abc import Generator
-from os import chdir
-from pathlib import Path
-from typing import Any
-
 import pytest
 from _pytest.monkeypatch import MonkeyPatch
 
 from tests import run, run_asserting_error
-
-
-@pytest.fixture
-def new_dir(tmp_path: Path) -> Generator[Path, Any, None]:
-    """Provide new test directory"""
-    current_directory = Path.cwd()
-    chdir(tmp_path)
-    yield tmp_path
-    # switch back to original directory
-    chdir(current_directory)
 
 
 def test_fails() -> None: